HTML schema and additional scripting
Implemented a way to create the cards in the deck, and to create a new card object with its associated values. Aces cannot yet adjust their values between 1 and 11.  Properly set up the HTML document and linked jquery, js and stylesheet. Sketched out some HTML elements to form the basic items we expect to display on the page.